make methods handler work in class style components
