+++ src/script-handlers/classMethodHandler.ts
@@ -1,9 +1,7 @@
 import * as bt from '@babel/types'
 import { NodePath } from 'ast-types'
-import { BlockTag } from 'src/utils/blockTags'
 import { Documentation, MethodDescriptor } from '../Documentation'
-import getDocblock from '../utils/getDocblock'
-import getDoclets, { DocBlockTags } from '../utils/getDoclets'
+import getMethodDescriptor from '../utils/getMethodDescriptor'
 
 export default function methodHandler(documentation: Documentation, path: NodePath) {
   if (bt.isClassDeclaration(path.node)) {
@@ -12,18 +10,11 @@ export default function methodHandler(documentation: Documentation, path: NodePa
       .get('body', 'body')
       .filter((a: NodePath) => (a.node as any).type === 'MethodDefinition')
 
-    allMethods.forEach((methodPath: NodePath<bt.Method>) => {
-      const docBlock = getDocblock(methodPath)
-      const jsDoc: DocBlockTags = docBlock ? getDoclets(docBlock) : { description: '', tags: [] }
-      const jsDocTags: BlockTag[] = jsDoc.tags ? jsDoc.tags : []
-      if (!jsDocTags.some((t) => t.title === 'public')) {
-        return
+    allMethods.forEach((methodPath: NodePath<bt.Property>) => {
+      const doc = getMethodDescriptor(methodPath)
+      if (doc) {
+        methods.push(doc)
       }
-      const name = methodPath.node.key.name
-      methods.push({
-        name,
-        description: jsDoc.description,
-      })
     })
     documentation.set('methods', methods)
   }
